Relocation modifyValueInObject
