+++ src/components/permissions/tests/rowLevelPolicyGeneration.test.ts
@@ -1,5 +1,5 @@
 import { getSqlConditionFromJSON } from '../helpersUtilities'
-import { compileRowLevelPolicies, modifyValueInObject } from '../rowLevelPolicyHelpers'
+import { compileRowLevelPolicies } from '../rowLevelPolicyHelpers'
 
 //. Tests to quickly check what is generated by row level policy helpers
 
@@ -60,19 +60,3 @@ rowLevelPolicyCompileTestData.forEach(
     })
   }
 )
-
-test(`Check modify in object`, () => {
-  return expect(
-    modifyValueInObject(
-      {
-        something: { $from: 'yow' },
-        somethingElse: { nested: [], alsoNested: { $from: 'hi' } },
-      },
-      (key, value) => key == '$from' && typeof value == 'string',
-      (value) => `private.${value}`
-    )
-  ).toEqual({
-    something: { $from: 'private.yow' },
-    somethingElse: { nested: [], alsoNested: { $from: 'private.hi' } },
-  })
-})
